+++ CHANGELOG.md
@@ -89,7 +89,7 @@
 ### Query-tee
 
 * [CHANGE] Query-Tee: Added `/api/v1/read` as a registered route. #13227
-* [CHANGE] Query-tee: Added cluster validation label configuration `-query-tee.client-cluster-validation.label`. If set, query-tee will set `X-Cluster` header before forwarding the requests to the backends. #13302
+* [CHANGE] Query-tee: Added cluster validation label configuration `-query-tee.client-cluster-validation.label`. If set, query-tee will set `X-Cluster` header before forwarding the request to both primary and secondary backends. #13302
 
 ## 3.0.0